requirements.txt
langchain
python-dotenv 
langchain_community
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
ipykernel
langchain_groq
streamlit 
langchain_core
duckduckgo-search
validators==0.28.1
youtube_transcript_api
unstructured
crewai
crewai_tools
pinecone
pinecone-client
pinecone-text
pinecone-notebooks

